Fix compilation problem on 64-bit machines.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@28817 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/paplayer/YMCodec/StSoundLibrary/XBMCYM.cpp
@@ -2,6 +2,7 @@
 #include "YmMusic.h"
 
 #include <stdio.h>
+#include <stdint.h>
 
 #ifdef __linux__
 #define __declspec(x)
@@ -17,7 +18,7 @@ extern "C"
     {
       ymMusicSetLoopMode(pMusic,YMFALSE);
       ymMusicPlay(pMusic);
-      return (int)pMusic;
+      return (intptr_t)pMusic;
     }
   
     ymMusicDestroy(pMusic);
